#170 fetch access token in requireTeam()
