+++ client/lib/common/src/util.js
@@ -1,0 +1,19 @@
+function requireTeam(flux) {
+    const ACTIONS = flux.getActions('user'),
+          STORE = flux.getStore('user');
+    if (!STORE.getUserTeams().length) {
+        let tokeninfo = STORE.getTokenInfo();
+        if (!tokeninfo.uid) {
+            return ACTIONS
+                    .fetchTokenInfo()
+                    .then(token => ACTIONS.fetchUserTeams(token.uid))
+                    .catch(() => ACTIONS.fetchAccessToken());
+        }
+        return ACTIONS.fetchUserTeams(tokeninfo.uid);
+    }
+    return Promise.resolve();
+}
+
+export {
+    requireTeam as requireTeam
+};
